+++ humbledb/__init__.py
@@ -18,7 +18,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 
 """
-__version__ = '5.5.0'
+__version__ = '5.5.1'
 
 
 # We only want to allow * imports for the most common classes. If you want